Delete tag and destination

+++ src/components/DestinationEdit/DestinationEdit.jsx
@@ -2,7 +2,13 @@ import React, { useState } from "react";
 import styles from "./DestinationEdit.module.css";
 import { useNavigate } from "react-router-dom";
 import { TbTrash } from "react-icons/tb";
-import { doc, deleteDoc, getFirestore } from "firebase/firestore";
+import {
+  doc,
+  deleteDoc,
+  getDocs,
+  collection,
+  getFirestore,
+} from "firebase/firestore";
 import appFirebase from "../../../credenciales";
 
 const db = getFirestore(appFirebase);
@@ -11,13 +17,30 @@ const DestinationEdit = ({ imagen, titulo, direccion, new: isNew, id }) => {
   const navigation = useNavigate();
   const [showConfirm, setShowConfirm] = useState(false);
   const [showSuccess, setShowSuccess] = useState(false);
+  const [alertMessage, setAlertMessage] = useState("");
 
   const handleDelete = async () => {
     try {
+      const reservationsRef = collection(db, `destinations/${id}/reservations`);
+      const reservationsSnapshot = await getDocs(reservationsRef);
+      const reservationsDocs = reservationsSnapshot.docs;
+      if (reservationsDocs.length > 1) {
+        setAlertMessage("No puedes eliminar este destino porque tiene reservas activas.");
+        setShowConfirm(false);
+        return;
+      }
+      const slugDoc = reservationsDocs[0].id;
+
+      if (slugDoc !== direccion) {
+        setAlertMessage("No puedes eliminar este destino porque tiene reservas activas.");
+        setShowConfirm(false);
+        return;
+      }
+
       await deleteDoc(doc(db, "destinations", id));
       setShowConfirm(false);
       setShowSuccess(true);
-  
+
       setTimeout(() => {
         setShowSuccess(false);
         navigation("/destinations-manage");
@@ -27,7 +50,6 @@ const DestinationEdit = ({ imagen, titulo, direccion, new: isNew, id }) => {
       alert("Hubo un error al eliminar el destino.");
     }
   };
-  
 
   return (
     <>
@@ -49,17 +71,29 @@ const DestinationEdit = ({ imagen, titulo, direccion, new: isNew, id }) => {
           </div>
         </div>
       )}
+
+      {alertMessage && (
+        <div className={styles.modalOverlay}>
+          <div className={styles.modalContent}>
+            <h3>{alertMessage}</h3>
+            <button
+              className={styles.cancelButton}
+              onClick={() => setAlertMessage("")}
+            >
+              Cerrar
+            </button>
+          </div>
+        </div>
+      )}
+
       {showSuccess && (
         <div className={styles.successMessage}>Destino eliminado con éxito</div>
       )}
+
       {!isNew ? (
-        <div className={`${styles.tarjetaDestino}`}>
-          <img
-            src={imagen}
-            alt={titulo}
-            className={`${styles.imagenDestino}`}
-          />
-          <div className={`${styles.contenidoDestino}`}>
+        <div className={styles.tarjetaDestino}>
+          <img src={imagen} alt={titulo} className={styles.imagenDestino} />
+          <div className={styles.contenidoDestino}>
             <h2 className={styles.tituloDestino}>{titulo}</h2>
             <button
               className={`btn-quaternary ${styles.btnEdit}`}
@@ -69,16 +103,16 @@ const DestinationEdit = ({ imagen, titulo, direccion, new: isNew, id }) => {
             </button>
           </div>
           <button
-            className={`${styles.trashButton}`}
+            className={styles.trashButton}
             onClick={() => setShowConfirm(true)}
           >
             <TbTrash />
           </button>
         </div>
       ) : (
-        <div className={`${styles.tarjetaDestino2}`}>
-          <div className={`${styles.imagenVerde}`}></div>
-          <div className={`${styles.contenidoDestino}`}>
+        <div className={styles.tarjetaDestino2}>
+          <div className={styles.imagenVerde}></div>
+          <div className={styles.contenidoDestino}>
             <h2 className={styles.tituloDestino}>{titulo}</h2>
             <button
               className={`btn-primary ${styles.btnEdit}`}
